refactor componentView to UIView
