+++ core/src/com/dragons/game/controller/playerController/PlayerController.java
@@ -4,8 +4,8 @@ import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.dragons.game.controller.gameWorld.GameObject;
 import com.dragons.game.controller.gameWorld.GameWorld;
-import com.dragons.game.view.componentViews.DropBombButtonView;
-import com.dragons.game.view.componentViews.JoystickView;
+import com.dragons.game.view.UIViews.DropBombButtonView;
+import com.dragons.game.view.UIViews.JoystickView;
 
 import net.dermetfan.gdx.assets.AnnotationAssetManager;
