+++ core/src/com/dragons/game/view/UIViews/LifeDisplayView.java
@@ -1,4 +1,4 @@
-package com.dragons.game.view.componentViews;
+package com.dragons.game.view.UIViews;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
@@ -6,7 +6,6 @@ import com.badlogic.gdx.math.Vector2;
 import com.dragons.game.model.IModel;
 import com.dragons.game.model.maps.GameMap;
 import com.dragons.game.model.players.IPlayer;
-import com.dragons.game.model.players.Player;
 import com.dragons.game.utilities.Constants;
 import com.dragons.game.view.IView;
